Moved code for better separation of concerns
